style: refactor sln and clean spaces and tabs

+++ 1_Presentation/Hub/src/Controllers/ImGuiController.cs
@@ -909,8 +909,8 @@ namespace Alis.App.Hub.Controllers
                 ImDrawListPtr cmdList = drawData.CmdListsRange[n];
 
                 // Upload vertex/index buffers
-                Gl.GlBufferData(BufferTarget.ArrayBuffer, (IntPtr) (cmdList.VtxBuffer.Size * drawVertSize), cmdList.VtxBuffer.Data, BufferUsageHint.StreamDraw);
-                Gl.GlBufferData(BufferTarget.ElementArrayBuffer, (IntPtr) (cmdList.IdxBuffer.Size * drawIdxSize), cmdList.IdxBuffer.Data, BufferUsageHint.StreamDraw);
+                Gl.GlBufferData(BufferTarget.ArrayBuffer, cmdList.VtxBuffer.Size * drawVertSize, cmdList.VtxBuffer.Data, BufferUsageHint.StreamDraw);
+                Gl.GlBufferData(BufferTarget.ElementArrayBuffer, cmdList.IdxBuffer.Size * drawIdxSize, cmdList.IdxBuffer.Data, BufferUsageHint.StreamDraw);
 
                 for (int cmdI = 0; cmdI < cmdList.CmdBuffer.Size; cmdI++)
                 {
@@ -988,9 +988,9 @@ namespace Alis.App.Hub.Controllers
             int uvOffset = 8; // Offset of Uv is 8 bytes from the start (after Pos)
             int colOffset = 16; // Offset of Col is 16 bytes from the start (after Pos and Uv)
 
-            Gl.VertexAttribPointer(SpaceWork.GlShader["Position"].Location, 2, VertexAttribPointerType.Float, false, drawVertSize, (IntPtr) posOffset);
-            Gl.VertexAttribPointer(SpaceWork.GlShader["UV"].Location, 2, VertexAttribPointerType.Float, false, drawVertSize, (IntPtr) uvOffset);
-            Gl.VertexAttribPointer(SpaceWork.GlShader["Color"].Location, 4, VertexAttribPointerType.UnsignedByte, true, drawVertSize, (IntPtr) colOffset);
+            Gl.VertexAttribPointer(SpaceWork.GlShader["Position"].Location, 2, VertexAttribPointerType.Float, false, drawVertSize, posOffset);
+            Gl.VertexAttribPointer(SpaceWork.GlShader["UV"].Location, 2, VertexAttribPointerType.Float, false, drawVertSize, uvOffset);
+            Gl.VertexAttribPointer(SpaceWork.GlShader["Color"].Location, 4, VertexAttribPointerType.UnsignedByte, true, drawVertSize, colOffset);
         }
     }
 }
